feat: add AS5600 component/sensor
The AS5600 is a magnetic rotary position encoder with a 12-bit resolution, allowing for 4096 positions/steps. There's been an open feature request for this for a couple years and a custom component implementation using a 3rd party library. This component uses the esphome native i2c implementation, not the 3rd party library.

resolves: esphome/feature-requests#1459
Signed-off-by: Branden Cash <203336+ammmze@users.noreply.github.com>