pn532: add support for polling FeliCa NFC cards
adds a component configuration that lets you set the card standard to
poll for. the default is to poll for ISO14443A, which is what the component
currently does.

the FeliCa standard is used in asian transit systems. if you, for example,
have an iphone loaded with a japanese Suica card with the express transit
option enabled, you can tap your phone to the reader without unlocking it to
trigger the `on_tag` event.