code ready for submitting wk2132 to esphome
I have started a new branch that incorporate the latest changes
in the code. Normally everything should be ok: I have reviewed the
code several times and tested that it execute correctly.
Note that this version of the code does not contain the GenUART
intermediate class (see branch wk2132old) as for now I am not
submitting the sc16is75x component.

On branch wk2132
Changes to be committed:
	modified:   CODEOWNERS
	new file:   esphome/components/wk2132/__init__.py
	new file:   esphome/components/wk2132/wk2132.cpp
	new file:   esphome/components/wk2132/wk2132.h
	modified:   tests/test1.yaml