Code fix #3 for PR - see details below
most of the changes are due to imptoper usage of damned const in C++!
among the changes one is tricky: in wk2132Component class I need a
counter to count the number of instances this counter is obviously
no a const so it has to initialized in the implementation file (.cpp).
this is done by int WK2132Component::counter{0}; but clang_tidy
complains because of a non constant variable accessible globally.
Therefore I had to use a new C++17 capability that allow to inline
and initialized a static attribute. If you know a better way let
me know. look at the following:
https://stackoverflow.com/questions/185844/how-to-initialize-private-static-members-in-c

On branch wk2132
Your branch is up to date with 'origin/wk2132'.

Changes to be committed:
	modified:   esphome/components/wk2132/wk2132.cpp
	modified:   esphome/components/wk2132/wk2132.h