Code modification number 3 - see details below
most of the changes are due to improper usage of damned const in C++!
among the changes one is tricky: in wk2132Component class I need a
counter to count the number of instances this counter is obviously
no a const so it has to initialized in the implementation file (.cpp).
this is done by int WK2132Component::counter{0}; but clang_tidy
complains because of a non constant variable accessible globally.
Therefore I had to use a new C++17 capability that allow to inline
and initialized a static attribute. If you know a better way let
me know. look at the following:
https://stackoverflow.com/questions/185844/how-to-initialize-private-static-members-in-c

On branch wk2132
Your branch is up to date with 'origin/wk2132'.

Changes to be committed:
	modified:   esphome/components/wk2132/wk2132.cpp
	modified:   esphome/components/wk2132/wk2132.h

Date:      Tue Aug 8 22:17:09 2023 +0200

interactive rebase in progress; onto a1a099a9
Last commands done (24 commands done):
   reword e9735727 modifications #2 as per PR request
   reword 426f3164 Code fix #3 for PR - see details below
Next commands to do (3 remaining commands):
   reword 7c93da1f Code fix number 4 for PR - see detail below
   reword 393d5301 Code fix number 4 - r=trying to fix again the i2s method!
You are currently editing a commit while rebasing branch 'wk2132' on 'a1a099a9'.

Changes to be committed:
	modified:   esphome/components/wk2132/wk2132.cpp
	modified:   esphome/components/wk2132/wk2132.h