Code modification number 7 - getting mysterious
The error found by clang-tidy are getting more and more
cryptic and mysterious. Hope for the best
Note: I have done rebase on the history to rewrite some of
the comments. Should be reflected now?

On branch wk2132
Your branch and 'origin/wk2132' have diverged,
and have 27 and 29 different commits each, respectively.
  (use "git pull" to merge the remote branch into yours)

Changes to be committed:
	modified:   esphome/components/wk2132/wk2132.cpp
	modified:   esphome/components/wk2132/wk2132.h