Code modification number 8 - Hope I understand
I still have one error, but now I think (?) I understand the messages
given by clang-tidy : it seems that is following the execution of a
call to REG_TO_STR_P1 with a value that should not happen in real
but that would result in an array overflow. Even though it should not
happen I have extended the array to make clang happy. Hope this
will work.

Please enter the commit message for your changes. Lines starting
with '#' will be ignored, and an empty message aborts the commit.

On branch wk2132
Your branch is up to date with 'origin/wk2132'.

Changes to be committed:
	modified:   esphome/components/wk2132/wk2132.cpp