Fix 24 bit signed integer parsing in sml parser
This fixes an issue when a smart meter sends a 24 bit signed integer value. As the bytes_to_int function only 8/16/32/64 bit values, it will default to 64 bits if presented with a 24 bit signed value. Without proper sign extension, this will be interpreted as a positive number. The bytes_to_uint function was modified to detect if a 24 bit negative integer is fed being converted and a proper sign extension is performed.