Implementation of a ring buffer (experimental)
This is a first prototype using a ring buffer. This is done to
be ready to transfer quickly bytes one by one without need to
make transactions on the i2c bus.
This is a very preliminary version just to test the concept

On branch wk2132-ring-b1
Changes to be committed:
	modified:   esphome/components/wk2132/wk2132.cpp
	modified:   esphome/components/wk2132/wk2132.h