Working with GenericUART rx buffering no tx buffering
This is a working version that uses the intermediate GenericUART
class. It is using the RingBuffer class for received bytes but
not for transmit bytes. Also the GenericUART requires a lot of
methods from the derived class (should be removed)
Also added the Doxygen configuration file for documentation.

On branch wk2132-genuart-b1
Changes to be committed:
	new file:   esphome/components/wk2132/gen_uart.cpp
	new file:   esphome/components/wk2132/gen_uart.h
	modified:   esphome/components/wk2132/wk2132.cpp
	new file:   esphome/components/wk2132/wk2132.doxygen
	modified:   esphome/components/wk2132/wk2132.h