Working with GenericUART and RingBuffer (clean)
This a a clean and tested version of wk2132 that use the
GenericUART class so we are ready to add more derived class
like sc16is75x type of component (i2c or spi!).
Now we also use the RingBuffer class to improve performance
especially when reading bytes one by one. It also provide a
better abstraction with the derived HW layer.

On branch wk2132-genuart-b1
Changes to be committed:
	modified:   esphome/components/wk2132/gen_uart.cpp
	modified:   esphome/components/wk2132/gen_uart.h
	modified:   esphome/components/wk2132/wk2132.cpp
	modified:   esphome/components/wk2132/wk2132.h