changed namespace gen_uart to wk2132
I do not like this modification but the script ci-custom
says: 'All integration C++ files should put all functions
in a separate namespace that matches the integration's name.'
therefore I changed the gen_uart namespece to wk2132!

On branch wk2132
Your branch is up to date with 'origin/wk2132'.

Changes to be committed:
	modified:   esphome/components/wk2132/gen_uart.cpp
	modified:   esphome/components/wk2132/gen_uart.h
	modified:   esphome/components/wk2132/wk2132.cpp
	modified:   esphome/components/wk2132/wk2132.h