transmit ring buffer gone again (see explanation)
Finally I removed the transmit ring buffer as it causes more problems
that it brings solution. Write() is never called. It is always
write_array that is used and in that case the ring buffer does
not help. in fact it is adding a delay ...

On branch wk2132-ring-b3
Your branch is up to date with 'origin/wk2132-ring-b3'.

Changes to be committed:
	modified:   esphome/components/wk2132/gen_uart.cpp
	modified:   esphome/components/wk2132/gen_uart.h
	modified:   esphome/components/wk2132/wk2132.cpp