Set RC Switch protocol TX carrier frequen to 38kHz
The RC Switch protocol can be used for both RF and IR remotes. RF
Remotes don't require generating a carrier signal, so, as pointed out in
the [remote transmitter configuration variables docs](https://esphome.io/components/remote_transmitter#configuration-variables),
that carrier frequency can be nullified by setting the duty cycle to
100%.

Since this protocol is also used with IR devices, a 50% duty cycle
should still result in a carrier frequency being generated. Therefore,
this sets the frequency to 38kHz in the protocol implementaion, allowing
the user to cancel it out in YAML config.

When this carrier frequency is left to 0, then even with the duty cycle
set to 50%, infrarer transmitters don't include the carrier, and thus
demodulators don't pick up the signal at all.

So to summarise, this one liner enables the use of the RC Switch
protocol for IR as well as RF.