Tuya Number: split "multiply" to a separate option
Commit 9d4f471 (PR #5108), part of ESPHome 2023.9.1, introduced a
multiplication factor for Tuya Number, to be used when converting from a
float (say, pH 5.2) to an integer-only data-point (say, 520).

However, that commit overloaded the "step" attribute, incorrectly
assuming that all stepped values are supposed to be multiplied. There
are devices in the wild that expect integers to be moved in steps, but
do not require any multiplication; for example, a dehumidifier that
	accepts an integer for target humidity, but only in steps of 5%.

With the aforementioned commit, it's impossible to correctly configure
such a device. Moreover, despite documented, it is counter-intuitive, as
"step" (CONF_STEP) now does something different compared to other
components.

To resolve this, revert "step" to its previous function, and add a
separate "multiply" attribute, using the standard CONF_MULTIPLY key.
This follows what other components do, such as for example the Modbus
Controller Number component, thereby providing an additional knob in a
standard way.

Fixes: esphome/issues#4928