Fix bug when gfonts specify both weight and italics
Without this the font spec is incorrectly generated as e.g. `:wght@1,700` which leads to an HTTP 400 error of "Invalid selector: Too many values in tuple". This patch adds the required leading `ital,` axis to the selector in this edge case. I've tested it with all permutations of weight and italics specified/not specified and it seems to reliably download fonts now.