Bump zeroconf to 0.123.0
changelog: https://github.com/python-zeroconf/python-zeroconf/compare/0.122.3...0.123.0

Before zeroconf 0.123.0, zeroconf would try to answer incoming questions
even if the service registry was empty which will always be the case for
the esphome cli since its only listening for services. This also meant
that it needless decoded all the known answers in each question packet
that it would never have an answer for. If the sevice registry is empty
it now skips the handle query path which significantly reduces background
cpu when the esphome cli is running.