Add a Binary Sensor Filter for state settling
This filter complements the DelayedOnOff filter, but instead of passing
the value after stabilization, this filter publishes a change in state
immediately and then waits to publish any additional changes until the
state remains stable for a given delay period.

The primary motivation for this filter is to reduce the latency incurred
by the Delayed* filter family while still debouncing the sensor, at the
cost of removing the noise filtering provided by the Delayed* filters.