Adding support for RDS-RGBW-02 DOP lights
This required significant re-architecting as the existing architecture
was built under the assumption one would be using 1- or 2-wire strings
that use simple RGB(W) arrays and shift subsequent data down the line.
The RDS-RDBW-02 and other data-over-power schemes require in-band
addressing and potentially other, more complex structuring, potentially
application-specific.

To enable this, the transformation of the color buffer to
`rmt_item32_t`s is facilitated by the `RMTView` abstraction. This lets
the component host swappable transformation code and can potentially allow
a user to implement their own schema in a config-included header file if
they so desire.