Remove compilation of test component routines
Now by default TEST_COMPONENT is not defined to reduce the size
of the wk2132 code. Compiling the component with the test routines
now requires the usage of build_flags in the yaml configuration
On branch wk2132
Your branch is up to date with 'origin/wk2132'.

Changes to be committed:
	modified:   esphome/components/wk2132_i2c/wk2132_i2c.h