Copy transform options from display to touchscreen
If the display has 'rotation' or 'transform' options, and the touchscreen
does not, create a 'transform' configuration for the touchscreen matching
the display's one.

This avoids duplicating the transformations when the display and the
touchscreen have the same (rotated) orientation, but can be easily overriden
for cases where the orientations differ.