Add `invert_position_report` to `tuya.cover`
Some Tuya cover MCUs for blinds persistently report a % open value for
the cover position, placing them at odds with ESPHome and HA controls.

This patch adds `invert_position_report` which simply flips the sense
of the relative position when used by the `tuya.cover` component so it
can be matched to whatever the device does.

This has been tested with a chip-swapped Zemismart `ZM25TQ (TY-W)`
successfully and gets the controls working consistently when used as
`device_class: blind` cover.