This is the  first working version with new structure
Now the code seems to work correctly. In the WK2132Component
class we create smart pointer to Register class to avoid
memory leakage. Now code seems to work OK for I2C bus interface
with the new classes adding SPI bus should be easy.
Still a lot of cleanup to do but code should be good for
tests

On branch wk2132_mix-b1
Changes to be committed:
	modified:   esphome/components/wk2132/wk2132.cpp
	modified:   esphome/components/wk2132/wk2132.h
	modified:   esphome/components/wk2132_i2c/wk2132_i2c.cpp
	modified:   esphome/components/wk2132_i2c/wk2132_i2c.h