wk2132_i2c works wk2132_spi does not work
cleanup code fix minor problem
fix write_fifo() wrong in spi version

Still does not work: Problem is that MISO signal output by
ESP21 is totally garbage. Transitions do not follow the CLK ?
Not sure if it is a problem of initialization ?
On branch wk2132_mix-b2
Your branch is up to date with 'origin/wk2132_mix-b2'.

Changes to be committed:
	modified:   esphome/components/wk2132/wk2132.h
	modified:   esphome/components/wk2132_i2c/wk2132_i2c.cpp
	modified:   esphome/components/wk2132_i2c/wk2132_i2c.h
	modified:   esphome/components/wk2132_spi/wk2132_spi.cpp
	modified:   esphome/components/wk2132_spi/wk2132_spi.h