Added GPIO pins version 2168_spi
lots of cleanup
added the GPIO support in the spi version.
Tested on a compatible 2132 (only 2 channel)
Cant test GPIO until I receive the test board
Should be able to use the wk_base comp in the wk2132
I will test it and push in the wk2132 pr if still time

On branch wk2168
Your branch is up to date with 'origin/wk2168'.

Changes to be committed:
	modified:   esphome/components/wk2168/__init__.py
	modified:   esphome/components/wk2168_spi/__init__.py
	modified:   esphome/components/wk2168_spi/wk2168_spi.h
	modified:   esphome/components/wk_base/wk_base.cpp
	modified:   esphome/components/wk_base/wk_base.h
	modified:   esphome/components/wk_base/wk_reg_def.h