New tft upload engine
This unifies the TFT upload engine between `arduino` and `esp-idf`, which increases consistency in the code and brings the advantages previously existing in one code to the other:

ESP-IDF now supports the persistent connection to http server.
Arduino now transfers with a stream, reducing the number of requests to the http server.