CSE7766: Fix energy calculation
Adding small float values to a large float value is not a good
idea. Due to the lack of accuracy, the small values may be completely
ignored.

This commit uses an unsigned integer to sum the cf pulses and calculate
the total energy from the total number of cf pulses.