Added new CMD and renamed CMD to match datasheet
Sometimes the AHT10 fails to recovery during firmware updates and a full power cycle is needed. By adding the soft reset command during setup as defined in the datasheet we can avoid the need to power cycle the device.

I renamed the Calibrate command to Initialize because that is the correct name in the datasheet. The AHT10 is calibrated at the factory, not by the Initialize command.