Port wifi_component_esp32_arduino to esp_netif
wifi_component_esp32_arduino is using the tcpip_adapter API from
ESP-IDF. tcpip_adapter, however, was deprecated in ESP-IDF 4.1, and
replaced with stubs that use the netif API internally. It was finally
removed in ESP-IDF 5.1, and is no longer available. We already depend on
ESP-IDF >= 4.1.

Replace all occurences of the tcpip_adapter API with the netif API
directly, borrowing code from the ESP-IDF WiFi component. tcpip_adapter
is effectively a wrapper over this code, so there are limited to no
functional changes.

This removes a dependency on a deprecated API, aligns the implementation
further with the ESP-IDF WiFi component, making some of the functions
there identical, allowing for future code sharing.

Note that this also resolves at least one more minor bug: the ability to
use IPv6 ULA addresses, something that was just not possible before due
to limitations of the tcpip_adapter API.