Nextion - Review types
This reviews types used in multiple places on Nextion component, which will result in a better memory management but, mostly, on preventing unexpected results when an invalid value is sent to Nextion display.
The function calls will now enforce the numeric ranges supported by Nextion or at least will manage this better than before.

This was part of #6192 which I'm breaking down.