Do not refresh sensors while updating
This prevents Nextion to request sensors/switches updates while uploading TFT, preventing adding those requests to the queue and preserving memory.
The device usually restarts after an update, so the queue will be cleaned up, so no reason to populate the queue during the upload.

This is part of #6192 which I'm breaking down.