Nextion - Review set_protocol_reparse_mode()
This rebuilds the `set_protocol_reparse_mode()` to use the `send_command_` call, avoiding direct calls to uart and standardizing the library as much as possible.
It also returns the status of sending the commands to Nextion, so it will return `false` if any of the commands fails.

This is part of the break down of #6192