Nextion - Use persistent http connection for TFT upload (Arduino)
TFT upload now transfers with a single stream, reducing the number of requests to the http server. In fact, this is similar to increasing the previously used chunk size to the file size, so all the data is requested to the http server in a single request, but then transferred to the UART on 4096 bytes chunks (as required by Nextion). If the display requests a different position, then the next request is the new http request (still in a persistent connection).
This will avoid a new http connection per chunk on Nextion when uploading TFT file, improving memory management and reliability of the whole process.
This is only for Arduino. The ESP-IDF equivalent is on #6576.
This is part of the breaking down of #6192