Nextion - Set alternative TFT update baud rate
The `upload_tft` function now accept as a parameters to define a target baud rate for the transfer (as Nextion supports the transfer in an alternative baud rate) . The parameters is optional, so this is not breaking existing implementations.

This function initiates the upload of a TFT file to the Nextion display. Users can specify a target baud rate for the transfer. If the provided baud rate is not supported by Nextion, the function defaults to using the current baud rate set for the display. If no baud rate is specified (or if 0 is passed), the current baud rate is used.

Supported baud rates are: 2400, 4800, 9600, 19200, 31250, 38400, 57600, 115200, 230400, 250000, 256000, 512000 and 921600. Selecting a baud rate supported by both the Nextion display and the host hardware is essential for ensuring a successful upload process.

@param baud_rate The desired baud rate for the TFT file transfer, specified as an unsigned 32-bit integer. If the specified baud rate is not supported, or if 0 is passed, the function will use the current baud rate. The default value is 0, which implies using the current baud rate.

This is part of the breaking down of #6192.