Merge upload tft engine in a single file
This is another try to merge the algorithms into a single file, as I've previously tried with #6192, but this time in a more simplified way.
The new file is the Arduino one with the modifications required by ESP-IDF. Although the #ifdef are not the best readable thing, it was just a couple of those new, so I believe this worth the change, as it makes way easier to maintain this code, reducing the need to repeat the changes on both files every time we want to make a change.