time_based_cover: rename `last_operation_` to `last_moving_operation_`
The previous implementation of the `last_operation` attribute was a misnomer: it represented the last moving operation of the cover, since it ignored all `COVER_OPERATION_IDLE` operations.
To keep compatibility with existing logic, an ad-hoc attribute `last_moving_operation` is introduced (and exposed via a getter method).
It has the same update logic as the previously named `last_operation` attribute.