uart: allow setting the UART id in final_validate_device_schema
uart.final_validate_device_schema() currently expects to validate a
device referenced by "uart_id". This is limiting, as it doesn't allow
components to define their own variable for defining the UART ID. (This
in turn may be useful in particular if a component supports or needs
multiple UARTs).

Allow defining the UART bus variable name, defaulting to CONF_UART_ID
for ease of use and backwards compatibility.