'uart' and 'improv_serial' need to understand non-UART logger configurations.
The 'uart' component assumed that the 'logger' component always used a
hardware UART, but that is not always true. If USB_CDC or USB_SERIAL_JTAG are
used for the 'logger', then no hardware UART is used.

The 'improv_serial' component has logic to deal with this situation,
but its use of preprocessor conditional logic duplicated the
conditional logic already in the 'logger' component; this PR changes
it to use the macros defined by the 'logger' component instead of
duplicating the logic.