Don't test for IPv6 addresses when min_ipv6_addr_count is 0
When IPv6 is enabled but `min_ipv6_addr_count` is set to zero (the default), we are testing whether a `uint8_t` value is greater or equal to zero here. This is of course always true, so the compiler emits a warning (due to `-Wtype-limits`).

Alter the `#if` slightly here to avoid this test when `min_ipv6_addr_count` is zero.