hydreon_rgxx: Fix parsing of data line
This fixes parsing of received data lines from a RG-15 sensor with
the hydreon_rgxx component.  In the situation that the "Acc"
sensor is not enabled in the yaml configuration, the data line
(which starts with Acc as the first reported value) would not
have been detected as an actual data line, and instead treated as
unparsable, with the values not filled into the sensors that are
configured.