hx711: Check for DOUT going high after a reading
The protocol for the HX711 is very simple, but it *is* possible to detect
that the device is absent and not just report zeroes as it does at the
moment:

 ﻿[D][hx711:033]: 'HX711 value': Got value 0

After a reading, when the 25th and subsequent pulses of the SCK line
are indicating the gain for the next reading, DOUT is supposed to be
high. Check for that, and discard the reading if it doesn't happen:

 [W][hx711:066]: HX711 DOUT pin not high after reading (data 0x0)!

The board which actually *does* have an HX711 attached also still works.

 [D][hx711:031]: 'HX711 value': Got value -1073110
 ﻿[D][sensor:094]: 'HX711 value': Sending state 66.13105 kg with 0 decimals of accuracy