[sml] Fixed crashing sml parser
The handling of the TL (type/length) field caused the parser
to die with certain edge-cases.
Basically the problem is that value fields count the TL bytes,
while lists do not count these.

Additionally there was a slight bug in the sign extension that
could cause some wrong values being read.

Signed-off-by: Nis Wechselberg <enbewe@enbewe.de>