Allow enable_ipv6 to be set to false on platforms which don't support it
I have a default configuration which uses IPv6, but some platforms only
support Legacy IP. It doesn't suffice just to override enable_ipv6=false
in their own config; we are never even allowed to *mention* IPv6.

Fix that, by adding a boolean_false validator. And put it first, so that
the error the user actually sees if they *do* set it to true, is the one
about it not being supported on that platform.