[libretiny] Report version 1.7.0 for 'dev' and 'latest'
Rather than reporting 0.0.0 for 'dev' and 'latest' branches, report
1.7.0 which is (currently) the truth. This will allow IPv6 to be used.

Like other frameworks, this will be perpetually out of date because we
don't check platformio/git and update it at runtime (although perhaps we
could at least have a script which updates it prior to making an ESPHome
release).

So unlike other frameworks, don't actually *clamp* to the version we
believe is the 'latest'. Explicitly set the source string to "libretiny"
without any version specifier so that we actually do get the latest.

This means that the version numbers being out of date is not a problem
because it's only really used for ESPHome to ensure that it has a new
enough version for certain features (like IPv6), so being newer than
we expect is just fine.