[font] Add support for "glyphsets"
For space and efficiency reasons, the font component currently defaults
to a small subset of glyphs, numbers, Latin alphabet characters and a
few symbols.

It also allows customizing this set of glyphs, by specifying a different
list of glyphs, or an additional set of glyphs. For speakers of
non-Latin based languages it is, however, non-trivial and error prone
for users to enumerate entire alphabets, especially when taking into
account all the permutations of characters and accents.

This adds a new "glyphsets" attribute, that leverages Google Fonts
Glyphset definitions, such as "GF_Latin_Core" and GF_Greek_Core", to
allow users to concisely include glyphs for their own script/language.

Signed-off-by: Faidon Liambotis <paravoid@debian.org>