Fix parsing of µs time periods in config
The correct SI prefix for micro is µ. In the 1980s where computers only
really had ASCII, some people got used to using 'u' instead, but it was
always technically wrong. In the 1990s, we had the legacy 8-bit character
sets so using anything outside the ASCII range was still problematic,
and people stuck with the 'u' approximation.

In the 21st century we have Unicode, the µ character is trivial to type
on many keyboards (it's just <Right-Alt + m> on my British keyboard).

I find myself typing time periods like '50µs' *correctly*, then getting
a build failure and having to adjust it to the 20th century version.