lps22: add a component
LPS22HB/HH is a barometric pressure sensor. It has pretty high accuracy,
is somewhat less prone to measuring weird stuff compared to competing
chips and is reasonably widely available for purchase.

This adds a very basic one-shot based implementation for interacting
with this sensor over an I2C bus. The chip also supports communication
over SPI, but I don't have the means to test such configuration (for my
PCB is baked for I2C) as well as measuring the conditions in a automatic
mode. These are left as future extensions to the basic implementation
seen here.