Use SOC_RMT_MEM_WORDS_PER_CHANNEL to set RMT_SYMBOL_CAPACITY
On ESP32-S3 mem_block_symbols must be at least 48. Set
RMT_SYMBOL_CAPACITY from SOC_RMT_MEM_WORDS_PER_CHANNEL which is the
minimum size accepted for TX or RX by the RMT driver; this is either 48
or 64 depending on the ESP32 flavour. Also add a static_assert() in case
any future flavours go below 40 symbols, which is the smallest number of
symbols we'd like to see.