opentherm: handle timeout during listen
This condition happens relatively frequently if another component hogs
the loop time, for example when the wifi component has to reconnect.
Occasionally an incoming message will be missed and the OpenTherm
component gets stuck listening forever.

This times out of the listening state after a reasonable amount of time
has passed.