[ina2xx_base] add doc for reset-on-boot disablement option (#5382)
* feat: add doc for reset-on-boot disablement option

Adds an entry in the documentation of the INA2xx sensor for the new
`reset_on_boot` option, which will allow the component to bypass
resetting the device state (and hence counters, etc.) on component
initialization. Setting this to `false` will allow the counters to
persist between ESP device reboots, while setting to `true` (the
default) will preserve the existing behaviour of resetting them.

* fix: add missing entry to SPI too

* fix: clarify wording for new option

* fix: improve reset_on_boot explanation line format

---------

Co-authored-by: Jonathan Swoboda <154711427+swoboda1337@users.noreply.github.com>