Fix parsing for modbus server mode
Previously all frames received in server mode were assumed to be client
commands to that server. If there are multiple servers on the bus then
responses from those other servers were not parsed correctly, resulting
in CRC errors and a string of other errors that follow.

This adds a flag to track client commands and server responses such that
the server can make an educated guess about what message it has
received.

It will attempt to re-parse the message in case CRC fails or a timeout
occurs to try to recover (in case there are two client commands in a row
to the same server).