[esphome OTA]Allow compilation on host platform
Allow compilation of esphome ota platform when using "host"
platform/framework. Real OTA funcitonality is not provided (or
emulated); this just allows compilation of configurations that
include esphome OTA in the host platform.

Sometimes some OTA hooks are included in some package files, that
are used simultaneously on the "real" device, and on the "host"
device while developing. This commit allows those configurations
to compile on "host", to minimize the differences in the configuration
yaml files.