Split ModbusDevice into ModbusServerDevice and ModbusClientDevice
For clients use pointers to return the result instead of an address
lookup.

This enables multiple client device configurations to talk to the same
device and retrieve the results separately. For example if one is
polling quickly and another slowly.

We still use address lookup for servers because it's initiated over the
wire.

Add ModbusDevice compatibility class for external components