Increase modbus default timeouts to reduce many common errors
Longer timeouts for modbus are generally always better if the
application allows it.
Assuming a response has timed out is bad when it comes later -
especially since it's a half duplex channel - so the next command
clashes with the "late" response of the previous command.

Here we're setting the timeout to 2s. This only affects clients that
don't respond. If they respond it has no effect.
We also set turnaround time to 600ms. This allows for slightly less than
100 messages per minute, which is plenty for folks sampling their
devices once per minute.

These values can always be changed by the user, but good defaults avoid
unnecessary bug reports.