Only clear parsed bytes when parse succeeds.
There is a situation where ModbusServer switches from server message
parsing to client message parsing (because a server failed to respond)
and the frame length may not match. In this case we don't want to hose
all of the bytes received in the buffer, but just the ones that parsed
correctly.