Clean up timeout calculation to make it reusable.
Reduce calls to millis() by keeping the last_receive_check timestamp
which can be used to calculate various timeouts (i.e. time since we last
checked for bytes, not the current time).

Simplify loop by having the parser check for the latest timeout value.
Timeout can change throughout the course of parsing as the length of the
buffer reduces - the last few bytes in the buffer could represent a
timeout.