Implement a one-shot queue for CommandItems
For components that utilise modbus_controller the previous pattern was to send the controller a CommandItem into a queue. We replicate that here since we need to keep the CommandItems around in memory until the response if received in order to fire the callbacks.
An alternative would be for components to keep the CommandItems locally and directly use send(). This avoids using the queue.