Add callback for not_sent, constructor, destructor
We add a callback for ModbusClient on_modbus_not_sent to allow us to track when a command doesn't actually go out (so we can clean it up from the queue)

Add convenience constructors to ModbusClientDevice, ModbusServerDevice
We also add a destructor for ModbusClientDevice that removes any references to itself from the modbus queue (via clear_rx_queue_for_device).
Provide a way to dequeue commands based on the device pointer (not just address)

Reset waiting_for_response as soon as we know we have a valid response.
This prevents unneccesary double clearing of waiting_for_response which can happen in the on_modbus_xx callbacks.