[ade7953] Apply voltage_gain setting to both channels (#12180)
The ADE7953 power sensor chip has one physical voltage sensor input, but
two registers that need to be programmed with the `voltage_gain`
setting.  We previously only applied the setting to channel A, which
meant that the power calculations for channel B were incorrect when a
non-default `voltage_gain` was used.  With this change, we also apply
`voltage_gain` to channel B.