Simplify ModbusServerDevice to return response or exception directly.
Since exactly one response is required from a server when it receives a request, and the address and function code need to match the request, there's no need to send them explicitly from the device - so we simply return the payload (or exception code) and the ServerHub will send the correct formatted response.

Now the ModbusServerDevice no longer needs to know about the parent (ServerHub). Much simpler and cleaner.